+++ @xen-orchestra/lite/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## **next**
 
-- Handle IPv6 hosts
+- Handle IPv6 hosts (PR [#8044](https://github.com/vatesfr/xen-orchestra/pull/8044))
+- [XOA deploy] Configure custom NTP servers (PR [#8059](https://github.com/vatesfr/xen-orchestra/pull/8059))
 
 ## **0.4.0** (2024-09-30)